requirements.txt
fastapi==0.91.0
numpy==1.21.6
paddleocr==2.6.1.2
paddlepaddle==2.4.1
Pillow==9.4.0
starlette==0.24.0
